chore: Move Note to the attribute package
