+++ src/main/java/seedu/loyaltylift/ui/customer/CustomerInfo.java
@@ -7,10 +7,11 @@ import javafx.scene.control.Label;
 import javafx.scene.control.ScrollPane;
 import javafx.scene.layout.AnchorPane;
 import javafx.scene.layout.VBox;
+import seedu.loyaltylift.model.attribute.Note;
 import seedu.loyaltylift.model.customer.Customer;
-import seedu.loyaltylift.model.customer.Note;
 import seedu.loyaltylift.model.order.Order;
 import seedu.loyaltylift.ui.Badge;
+import seedu.loyaltylift.ui.NotePanel;
 import seedu.loyaltylift.ui.UiPart;
 
 /**
@@ -48,7 +49,7 @@ public class CustomerInfo extends UiPart<ScrollPane> {
         // Note
         Note note = customer.getNote();
         if (!note.value.trim().isBlank()) {
-            CustomerNote customerNote = new CustomerNote(note);
+            NotePanel customerNote = new NotePanel(note);
             insertSection("Note", customerNote.getRoot());
         }
