+++ src/main/resources/view/NotePanel.fxml
@@ -3,6 +3,6 @@
 <?import javafx.scene.text.Text?>
 <?import javafx.scene.layout.VBox?>
 
-<VBox spacing="5.0" styleClass="base" stylesheets="@CustomerSection.css" xmlns="http://javafx.com/javafx/11" xmlns:fx="http://javafx.com/fxml/1">
+<VBox spacing="5.0" styleClass="base" stylesheets="@NotePanel.css" xmlns="http://javafx.com/javafx/11" xmlns:fx="http://javafx.com/fxml/1">
    <Text fx:id="note" text="\$note" />
 </VBox>
